Fix redundant stat in account and container server
Multiple stat() calls were made while serving GET requests for container
and account. This removes those calls and can be easily verified using
strace.

There is room for further refactoring of code to simplify it. This will
be addressed as a separate change to keep things simple in this patch.

Change-Id: Ief457ff869c58519e9dbeb4ef13797185f536673
Signed-off-by: Prashanth Pai <ppai@redhat.com>
Reviewed-on: http://review.gluster.org/14543
Reviewed-by: Thiago da Silva <thiago@redhat.com>
Tested-by: Thiago da Silva <thiago@redhat.com>